:bug: 글쓰기가 동작하지 않던 버그 수정
단, 새로 추가된 글이 제대로 동작하는 거는 보장하지 않습니다
