+++ src/pages/Home/sections/Community/CommunityPreviewList.jsx
@@ -19,12 +19,12 @@ function CommunityPreviewListRenderer({ resource }) {
       {boardList.map(({ board_no, board_title, board_type, board_content }) => {
         const boardGroupName = board_type === 1 ? "consult" : "review";
         return (
-          // <Link to={`/community/${boardGroupName}/${board_no}`} key={board_no}>
+          <Link to={`/community/${boardGroupName}/${board_no}`} key={board_no}>
             <CommunityPreviewArticle
               title={board_title}
               content={board_content}
             />
-          // </Link>
+          </Link>
         );
       })}
     </Wrapper>
